[Touch to Search] Fixes Peek Promo regression.
This a regression caused by:
https://codereview.chromium.org/1716653002/

Prior to the regression, when Touch to Search Bar was
peeking and the Peek Promo was showing, touching on the
base page and dragging would cause the base page to scroll,
causing the Bar to be dismissed.

The regression broke this behavior and now when the Peek
Promo is showing, trying to scroll the base page will
actually cause the Touch to Search Panel to be dragged.
This is bad because scrolling is one of the most common
ways of dismissing the Panel.

The problem was that the OverlayPanel was miscalculating
the height of the Bar, not taking into consideration the
height of the Peek Promo. This CL fixes this behavior.

BUG=613235

Review-Url: https://codereview.chromium.org/2006653004
Cr-Commit-Position: refs/heads/master@{#395992}