Don't reset the codec state for a flush; this kills the frames.
Causing a reset after flush completes will invalidate all the
vended frames, don't do this. Instead only reset when a reset
should actually occur.

Instead, since this reset is required on JellyBean devices, we
issue the reset upon receipt of the next Decode() call. This
allows the more frequent EOS cases to not trash the last frame
on these devices will config changes will work as before.

A followup CL will move the texture matrix to the shared state
so that we don't accidentally apply the wrong matrix after the
reset case.

BUG=613608
TEST=no more random blue frames.

Review-Url: https://codereview.chromium.org/2000833003
Cr-Commit-Position: refs/heads/master@{#395995}