Make it safe for the DownloadProtectionService to destroy an active PPAPIDownloadRequest.
When an active PPAPIDownloadRequest is destroyed, it notifies the owning
DownloadProtectionService that the request can now be removed from the
pending request list. If the server itself is destroying the request,
this can end up with the service attempting to destroy the request
twice.

Instead, prevent the PPAPIDownloadRequest from notifying its owner if
the request is being destroyed. The ownership relationship implies that
the owner already knows that the request is being destroyed.

R=nparker@chromium.org
BUG=614936

Review-Url: https://codereview.chromium.org/2012243002
Cr-Commit-Position: refs/heads/master@{#396219}