GLContextGLX: try all GL versions from the highest.
Also try all ES versions from the highest. This change will allow
Chrome to create a GL context higher than 3.0 on Mesa and will in turn
allow WebGL2 to run on Intel Linux when using the native GL driver.

This patch is a direct port of the logic in ANGLE's DisplayGLX

BUG=598902

Review-Url: https://codereview.chromium.org/1999543003
Cr-Commit-Position: refs/heads/master@{#396243}