cc: Add software image decodes usage UMAs.
This patch adds two new UMA stats.

Renderer4.SoftwareImageDecodeState
==================================
This is a general lifetime of an image decode stat, logged at decoded
image destruction. Each image will have one of the following states:
- Wasted: this image was never used and was not ever relocked
- Used: this image was used once and never relocked
- Used, relock failed: similar to used, but a relock was attempted and
  failed.
- Wasted, relocked: this image was never used and we've relocked it at
  least once.
- Used, relocked: this image was used and relocked.

Renderer4.SoftwareImageDecodeState.FirstLockWasted
==================================================
This logs whether we used the image when we first created it in a lock
state. That is, if we unlocked the image before using it, then the
first lock is wasted regardless of whether the image was used later
after relocking.

R=enne, ericrk
CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel

Review-Url: https://codereview.chromium.org/2009913003
Cr-Commit-Position: refs/heads/master@{#396297}