Retire setGradientSpaceTransform, setPatternSpaceTransform
Most gradient/pattern clients always reset the transform before
applying to a paint, while few (CanvasPattern) actually rely on the
transform state being persistent.  This yields confusing semantics
where it's unclear whether it is safe to whack the state.

The CL removes the transform from the gradient/pattern state, and makes
it an argument to applyToPaint().  This simplifies the gradient/pattern
impl and clarifies that clients are responsible for tracking the local
matrix.  The local matrix is still used as part of the shader cache key
- but since SkShaders already track their local matrix, there is no
need for explicit tracking in Gradient/Pattern.

R=fs@opera.com,junov@chromium.org
BUG=614368

Review-Url: https://codereview.chromium.org/2007553002
Cr-Commit-Position: refs/heads/master@{#396557}