Use only DIPs size in ClientResolution.
ClientResolution message contains DIPs and device size fields. DIPs
size was marked as legacy, but it was actually used by the host. It's
best to use DIPs for consistency with other messages, particularly
VideoTrackLayout. marked device-pixel size as deprecated and the host
now works when width/height fields are not specified.

Review-Url: https://codereview.chromium.org/2021773002
Cr-Commit-Position: refs/heads/master@{#397102}