Add a FakeImeFeatureDelegate to the Linux client.
The BlimpClientSession creates an ImeFeature which requires that a
Delegate be provided before the first IME message is received. Since the
Linux client is for developer testing only, we add a fake delegate which
just logs the IME show/hide requests.

This CL also corrects some BlimpMessageProcessor::ProcessMessage()
overrides to be public.

BUG=615584

Review-Url: https://codereview.chromium.org/2023613002
Cr-Commit-Position: refs/heads/master@{#397478}